feat: switch to react lazy for chunky zxcvb (so called "rap library")
