+++ packages/wallets/src/components/Base/WalletPasswordField/WalletPasswordField.tsx
@@ -1,17 +1,31 @@
 import React, { useCallback, useState } from 'react';
-import { passwordErrorMessage } from '../../../constants/password';
-import { Score, validatePassword, validPassword } from '../../../utils/password';
+import { Score, calculateScore, validPassword, isPasswordValid, passwordKeys } from '../../../utils/password';
 import { WalletTextField } from '../WalletTextField';
-import { WalletTextFieldProps } from '../WalletTextField/WalletTextField';
 import PasswordMeter from './PasswordMeter';
 import PasswordViewerIcon from './PasswordViewerIcon';
 import './WalletPasswordField.scss';
+import { passwordErrorMessage, passwordRegex, warningMessages } from '../../../constants/password';
+import { zxcvbn, zxcvbnOptions } from '@zxcvbn-ts/core';
+import { dictionary } from '@zxcvbn-ts/language-common';
+import { WalletPasswordFieldProps } from '../WalletPasswordFieldLazy/WalletPasswordFieldLazy';
 
-interface WalletPasswordFieldProps extends WalletTextFieldProps {
-    password: string;
-    passwordError?: boolean;
-    shouldDisablePasswordMeter?: boolean;
-}
+export const validatePassword = (password: string) => {
+    const score = calculateScore(password);
+    let errorMessage = '';
+
+    const options = { dictionary: { ...dictionary } };
+    zxcvbnOptions.setOptions(options);
+
+    const { feedback } = zxcvbn(password);
+    if (!passwordRegex.isLengthValid.test(password)) {
+        errorMessage = passwordErrorMessage.invalidLength;
+    } else if (!isPasswordValid(password)) {
+        errorMessage = passwordErrorMessage.missingCharacter;
+    } else {
+        errorMessage = warningMessages[feedback.warning as passwordKeys] ?? '';
+    }
+    return { errorMessage, score };
+};
 
 const WalletPasswordField: React.FC<WalletPasswordFieldProps> = ({
     autoComplete,
