--- packages/wallets/src/components/Base/WalletPasswordField/index.ts
@@ -1,1 +1,0 @@
-export { default as WalletPasswordField } from './WalletPasswordField';
