+++ packages/wallets/src/components/Base/WalletPasswordFieldLazy/index.ts
@@ -1,0 +1,1 @@
+export { default as WalletPasswordFieldLazy } from './WalletPasswordFieldLazy';
