+++ packages/wallets/src/components/Base/index.ts
@@ -13,6 +13,6 @@ export * from './WalletButton';
 export * from './WalletButtonGroup';
 export * from './WalletClipboard';
 export * from './WalletDropdown';
-export * from './WalletPasswordField';
+export * from './WalletPasswordFieldLazy';
 export * from './WalletText';
 export * from './WalletTextField';
