+++ packages/wallets/src/features/cfd/screens/ChangePassword/InvestorPassword/MT5ChangeInvestorPasswordInputsScreen.tsx
@@ -1,11 +1,12 @@
 import React from 'react';
 import { Formik } from 'formik';
 import { useTradingPlatformInvestorPasswordChange } from '@deriv/api';
-import { WalletButton, WalletPasswordField, WalletsActionScreen, WalletText } from '../../../../../components';
+import { WalletButton, WalletsActionScreen, WalletText } from '../../../../../components';
 import { useModal } from '../../../../../components/ModalProvider';
 import useDevice from '../../../../../hooks/useDevice';
 import { validPassword } from '../../../../../utils/password';
 import { PlatformDetails } from '../../../constants';
+import { WalletPasswordFieldLazy } from '../../../../../components/Base';
 
 type TFormInitialValues = {
     currentPassword: string;
@@ -65,14 +66,14 @@ const MT5ChangeInvestorPasswordInputsScreen: React.FC<TProps> = ({ sendEmail, se
                     {({ handleChange, handleSubmit, values }) => (
                         <form className='wallets-change-investor-password-screens__form' onSubmit={handleSubmit}>
                             <div className='wallets-change-investor-password-screens__form-fields'>
-                                <WalletPasswordField
+                                <WalletPasswordFieldLazy
                                     autoComplete='current-password'
                                     label='Current investor password'
                                     name='currentPassword'
                                     onChange={handleChange}
                                     password={values.currentPassword}
                                 />
-                                <WalletPasswordField
+                                <WalletPasswordFieldLazy
                                     autoComplete='new-password'
                                     label='New investor password'
                                     name='newPassword'
